response_pipeline/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Azure SDK
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.10.0

# LLM Support
tiktoken==0.5.2

# Redis (removed redis-om due to pydantic conflict)
redis==5.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aioodbc==0.5.0
pyodbc==5.0.1

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
rich==13.7.0
numpy==2.2.6

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# Type checking
types-redis==4.6.0.20240106
types-python-dateutil==2.8.19.20240106